Test:WIP:Refactoring tests and parametrize.yaml.
