+++ tests/mock/application/fsm_tools/locales/_i18n_tools/failed-i18n-tools.yaml
@@ -1,0 +1,83 @@
+application:
+  authors:
+    7d097ac4-ba77-4333-a63f-76d48a75b38c:
+      email: john@doe.com
+      first_name: John
+      languages:
+      - fr
+      - en
+      last_name: Doe
+      url: ''
+    baf52b25-d1ed-4651-a3e6-273850901cf0:
+      email: jeanne@doe.com
+      first_name: Jeanne
+      languages:
+      - fr
+      - en
+      last_name: Doe
+      url: ''
+  domains:
+    fsm_tools:
+    - usage
+    - model
+    fsm_tools/lba:
+    - information
+    - error
+    fsm_tools/turing:
+    - information
+    - error
+    django-fsm_tools:
+      - usage
+      - information
+      - error
+    django-fsm_tools/context:
+      - usage
+      - error
+      - output
+  language:
+    fallback: en
+    hierarchy:
+      en:
+      - en-US
+      - en-GB
+      fr:
+      - fr-FR
+    source: en
+  paths:
+    config: /home/fabricemoutte/Personnel/dev/python/i18n/tests/mock/application/fsm_tools/locales/_i18n_tools
+    modules:
+    - fsm_tools
+    - fsm_tools/turing
+    - fsm_tools/lba
+    - django-fsm_tools
+    - django-fsm_tools/context
+    repository: /home/fabricemoutte/Personnel/dev/python/i18n/tests/mock/application
+    root: /home/fabricemoutte/Personnel/dev/python/i18n/tests/mock/application
+    backup: /home/fabricemoutte/Personnel/dev/python/i18n/tests/mock/application/fsm_tools/locales/_i18n_tools/backup
+    settings: i18n-tools.yaml
+  translators:
+    GoogleTranslate:
+      details:
+        name: GoogleTranslate
+        status: free
+        url: https://translate.google.com
+      pricing:
+        cost_per_translation: null
+        payment_plan: null
+      technical:
+        api:
+          key: your_api_key
+          key_expiration: '2025-12-31'
+          request_limit: 10000
+        performance:
+          max_text_size: 5000
+          priority: 1
+          success_rate: 99.5
+  details:
+    description: This project is a study of implementation of Chomsky's formal language
+      and usage of finite state machine
+    name: fsm-tools
+    summary: A toolbox for Chomsky's formal language hierarchy
+    version: 0.0.1
+
+configuration: application
